+++ src/ImportDefinitionsBundle/Exporter/ExporterInterface.php
@@ -12,11 +12,11 @@
  * @license    https://github.com/w-vision/ImportDefinitions/blob/master/gpl-3.0.txt GNU General Public License version 3 (GPLv3)
  */
 
-namespace ImportDefinitionsBundle\Export;
+namespace ImportDefinitionsBundle\Exporter;
 
 use ImportDefinitionsBundle\Model\ExportDefinitionInterface;
 
-interface ExportInterface
+interface ExporterInterface
 {
     /**
      * @param ExportDefinitionInterface $definition
